Organized files into respective folders
